removes geo pipeline from intercensal folder

+++ etl/inegi_intercensal_census/population_pipeline.py
@@ -49,7 +49,7 @@ class PopulationPipeline(EasyPipeline):
 
         dtype = {
             'sex':          'UInt8',
-            'mun_id':       'UInt32',
+            'mun_id':       'UInt16',
             'population':   'UInt32',
         }
